Added speech recognition functionality
Instantiated the speech recognition engine
Added the grammar with the vocabulary list from the grammar.txt file dynamically.
Added UDP socket connection snippet which acts as inter process communication mechanism. This program acts as the server which will detect speech inputs from the user and sends the detected word to the <IPAddress:192.168.0.20>/<port:26000>